Moved implementation of heavy-weight functions in WindowTree.cpp
