+++ source/wm/WindowTree.cpp
@@ -1,1 +1,72 @@
 #include "wm/WindowTree.hpp"
+
+namespace wm
+{
+  WindowTree::WindowTree(WindowData &&screen)
+    : freeList(nullNode)
+  {
+    nodes.emplace_back(WindowNode{nullNode, nullNode, nullNode, std::move(screen)});
+  }
+
+  WindowTree::WindowTree(WindowData const &screen)
+    : freeList(nullNode)
+  {
+    nodes.emplace_back(WindowNode{nullNode, nullNode, nullNode, screen});
+  }
+
+  WindowNodeIndex WindowTree::allocateIndex()
+  {
+    if (freeList == nullNode)
+      {
+	WindowNodeIndex result(static_cast<uint16_t>(nodes.size()));
+
+	nodes.emplace_back();
+	return result;
+      }
+    else
+      {
+	WindowNodeIndex result(freeList);
+
+	freeList = getSibling(freeList);
+	return result;
+      }
+  }
+
+  void WindowTree::removeIndex(WindowNodeIndex index) noexcept
+  {
+    WindowNodeIndex child(getFirstChild(getParent(index)));
+
+    if (child == index)
+      getNode(getParent(index)).firstChild = getSibling(index);
+    else
+      {
+	while (getSibling(child) != index)
+	  child = getSibling(child);
+	getNode(child).nextSibling = getSibling(index);
+      }
+    getNode(index).nextSibling = freeList;
+    freeList = index;
+  }
+
+  WindowNodeIndex WindowTree::addChild(WindowNodeIndex parent)
+  {
+    WindowNodeIndex result(allocateIndex());
+
+    getNode(result).parent = parent;
+    getNode(result).nextSibling = getFirstChild(parent);
+    getNode(result).firstChild = nullNode;
+    getNode(parent).firstChild = result;
+    return result;
+  }
+
+  WindowNodeIndex WindowTree::addChildAfter(WindowNodeIndex parent, WindowNodeIndex index)
+  {
+    WindowNodeIndex result(allocateIndex());
+
+    getNode(result).parent = parent;
+    getNode(result).nextSibling = getNode(index).nextSibling;
+    getNode(result).firstChild = nullNode;
+    getNode(index).nextSibling = result;
+    return result;
+  }
+}
